+++ ethereum/core/src/main/java/tech/pegasys/pantheon/ethereum/mainnet/ProtocolSchedule.java
@@ -12,9 +12,11 @@
  */
 package tech.pegasys.pantheon.ethereum.mainnet;
 
+import java.math.BigInteger;
+
 public interface ProtocolSchedule<C> {
 
   ProtocolSpec<C> getByBlockNumber(long number);
 
-  int getChainId();
+  BigInteger getChainId();
 }
